test: add apache license in the test files.
