+++ test/runTest/actions/dataZoom-rainfall-connect.json
@@ -1,1 +1,1 @@
-[{"name":"Action 1","ops":[{"type":"mousedown","time":468,"x":138,"y":131},{"type":"mousemove","time":593,"x":140,"y":131},{"type":"mousemove","time":794,"x":274,"y":131},{"type":"mousemove","time":1081,"x":428,"y":124},{"type":"mouseup","time":1255,"x":428,"y":124},{"time":1256,"delay":400,"type":"screenshot-auto"},{"type":"mousemove","time":1288,"x":428,"y":124},{"type":"mousemove","time":1377,"x":426,"y":124},{"type":"mousemove","time":1577,"x":206,"y":119},{"type":"mousedown","time":1720,"x":184,"y":120},{"type":"mousemove","time":1786,"x":184,"y":120},{"type":"mousemove","time":1991,"x":407,"y":116},{"type":"mousemove","time":2191,"x":443,"y":115},{"type":"mouseup","time":2605,"x":443,"y":115},{"time":2606,"delay":400,"type":"screenshot-auto"},{"type":"mousemove","time":2995,"x":442,"y":116},{"type":"mousemove","time":3195,"x":347,"y":274},{"type":"mousemove","time":3395,"x":301,"y":351},{"type":"mousemove","time":3595,"x":328,"y":359},{"type":"mousedown","time":3679,"x":328,"y":359},{"type":"mousemove","time":3821,"x":324,"y":360},{"type":"mousemove","time":4037,"x":136,"y":364},{"type":"mousemove","time":4251,"x":100,"y":363},{"type":"mouseup","time":4330,"x":100,"y":363},{"time":4331,"delay":400,"type":"screenshot-auto"},{"type":"mousemove","time":4463,"x":188,"y":358},{"type":"mousemove","time":4663,"x":330,"y":355},{"type":"mousedown","time":4774,"x":329,"y":355},{"type":"mousemove","time":4870,"x":316,"y":359},{"type":"mousemove","time":5079,"x":132,"y":372},{"type":"mousemove","time":5301,"x":110,"y":374},{"type":"mouseup","time":5324,"x":110,"y":374},{"time":5325,"delay":400,"type":"screenshot-auto"},{"type":"mousemove","time":5683,"x":114,"y":375},{"type":"mousemove","time":5884,"x":171,"y":489},{"type":"mousemove","time":6083,"x":173,"y":506},{"type":"mousemove","time":6292,"x":175,"y":511},{"type":"mousemove","time":6349,"x":175,"y":511},{"type":"mousemove","time":6557,"x":173,"y":514},{"type":"mousemove","time":6768,"x":170,"y":521},{"type":"mousedown","time":6943,"x":169,"y":522},{"type":"mousemove","time":6976,"x":169,"y":522},{"type":"mousemove","time":7199,"x":214,"y":528},{"type":"mousemove","time":7403,"x":298,"y":528},{"type":"mouseup","time":7780,"x":298,"y":528},{"time":7781,"delay":400,"type":"screenshot-auto"},{"type":"mousemove","time":7951,"x":298,"y":527},{"type":"mousemove","time":8152,"x":177,"y":506},{"type":"mousemove","time":8361,"x":170,"y":500},{"type":"mousedown","time":8564,"x":170,"y":500},{"type":"mouseup","time":8646,"x":170,"y":500},{"time":8647,"delay":400,"type":"screenshot-auto"},{"type":"mousedown","time":9279,"x":170,"y":500},{"type":"mouseup","time":9389,"x":170,"y":500},{"time":9390,"delay":400,"type":"screenshot-auto"},{"type":"mousemove","time":9676,"x":171,"y":498},{"type":"mousemove","time":9877,"x":284,"y":120},{"type":"mousemove","time":10078,"x":299,"y":32},{"type":"mousemove","time":10289,"x":285,"y":1},{"type":"mousemove","time":10644,"x":285,"y":1},{"type":"mousemove","time":10854,"x":291,"y":6},{"type":"mousedown","time":11106,"x":291,"y":6},{"type":"mouseup","time":11189,"x":291,"y":6},{"time":11190,"delay":400,"type":"screenshot-auto"},{"type":"mousedown","time":11793,"x":291,"y":6},{"type":"mouseup","time":11890,"x":291,"y":6},{"time":11891,"delay":400,"type":"screenshot-auto"}],"scrollY":0,"scrollX":0,"timestamp":1568037648479}]
+[{"name":"Action 1","ops":[{"type":"mousedown","time":468,"x":138,"y":131},{"type":"mousemove","time":593,"x":140,"y":131},{"type":"mousemove","time":794,"x":274,"y":131},{"type":"mousemove","time":1081,"x":428,"y":124},{"type":"mouseup","time":1255,"x":428,"y":124},{"time":1256,"delay":800,"type":"screenshot-auto"},{"type":"mousemove","time":1288,"x":428,"y":124},{"type":"mousemove","time":1377,"x":426,"y":124},{"type":"mousemove","time":1577,"x":206,"y":119},{"type":"mousedown","time":1720,"x":184,"y":120},{"type":"mousemove","time":1786,"x":184,"y":120},{"type":"mousemove","time":1991,"x":407,"y":116},{"type":"mousemove","time":2191,"x":443,"y":115},{"type":"mouseup","time":2605,"x":443,"y":115},{"time":2606,"delay":800,"type":"screenshot-auto"},{"type":"mousemove","time":2995,"x":442,"y":116},{"type":"mousemove","time":3195,"x":347,"y":274},{"type":"mousemove","time":3395,"x":301,"y":351},{"type":"mousemove","time":3595,"x":328,"y":359},{"type":"mousedown","time":3679,"x":328,"y":359},{"type":"mousemove","time":3821,"x":324,"y":360},{"type":"mousemove","time":4037,"x":136,"y":364},{"type":"mousemove","time":4251,"x":100,"y":363},{"type":"mouseup","time":4330,"x":100,"y":363},{"time":4331,"delay":800,"type":"screenshot-auto"},{"type":"mousemove","time":4463,"x":188,"y":358},{"type":"mousemove","time":4663,"x":330,"y":355},{"type":"mousedown","time":4774,"x":329,"y":355},{"type":"mousemove","time":4870,"x":316,"y":359},{"type":"mousemove","time":5079,"x":132,"y":372},{"type":"mousemove","time":5301,"x":110,"y":374},{"type":"mouseup","time":5324,"x":110,"y":374},{"time":5325,"delay":800,"type":"screenshot-auto"},{"type":"mousemove","time":5683,"x":114,"y":375},{"type":"mousemove","time":5884,"x":171,"y":489},{"type":"mousemove","time":6083,"x":173,"y":506},{"type":"mousemove","time":6292,"x":175,"y":511},{"type":"mousemove","time":6349,"x":175,"y":511},{"type":"mousemove","time":6557,"x":173,"y":514},{"type":"mousemove","time":6768,"x":170,"y":521},{"type":"mousedown","time":6943,"x":169,"y":522},{"type":"mousemove","time":6976,"x":169,"y":522},{"type":"mousemove","time":7199,"x":214,"y":528},{"type":"mousemove","time":7403,"x":298,"y":528},{"type":"mouseup","time":7780,"x":298,"y":528},{"time":7781,"delay":800,"type":"screenshot-auto"},{"type":"mousemove","time":7951,"x":298,"y":527},{"type":"mousemove","time":8152,"x":177,"y":506},{"type":"mousemove","time":8361,"x":170,"y":500},{"type":"mousedown","time":8564,"x":170,"y":500},{"type":"mouseup","time":8646,"x":170,"y":500},{"time":8647,"delay":800,"type":"screenshot-auto"},{"type":"mousedown","time":9279,"x":170,"y":500},{"type":"mouseup","time":9389,"x":170,"y":500},{"time":9390,"delay":800,"type":"screenshot-auto"},{"type":"mousemove","time":9676,"x":171,"y":498},{"type":"mousemove","time":9877,"x":284,"y":120},{"type":"mousemove","time":10078,"x":299,"y":32},{"type":"mousemove","time":10289,"x":285,"y":1},{"type":"mousemove","time":10644,"x":285,"y":1},{"type":"mousemove","time":10854,"x":291,"y":6},{"type":"mousedown","time":11106,"x":291,"y":6},{"type":"mouseup","time":11189,"x":291,"y":6},{"time":11190,"delay":800,"type":"screenshot-auto"},{"type":"mousedown","time":11793,"x":291,"y":6},{"type":"mouseup","time":11890,"x":291,"y":6},{"time":11891,"delay":800,"type":"screenshot-auto"}],"scrollY":0,"scrollX":0,"timestamp":1568037648479}]
